pyproject.toml
[project]
name = "bid-announce-system"
version = "3.0.0"
description = "中标结果公示系统 - 数据采集与展示"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "httpx>=0.27.0",
]

[project.scripts]
bid-announce = "scripts.start_server:main"

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = ["src/api/app_v2.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
